Rackspace Cloud Files action
First version of a Rackspace Cloud Files action. It can upload file(s) to the configured Rackspace Cloud Files account. It also needs a container name which will be configured when the first file is uploaded. In order to reconfigure the container you can click on the action. Files CDN URLs are copied to clipboard after upload.