Switched action to Ruby 1.8. Other updates
Refactored the action from camelCase to snake_case. Not handling Text anymore, as the Slack API doesn't allow you to post messages into a channel as yourself, only as a bot. Moved the action to Ruby 1.8 and changed some require statements so the multipost gem is seen in the lib folder.